Changed data type of SubGroupSize from int to uint64 (#1927)
* Changed data type of SubGroupSize from int to uint because int can't be RLP encoded and it makes to use unnecessary json marshal/unmarshaling

* Fix test code to follow data type change of subsize

+++ consensus/istanbul/backend.go
@@ -78,5 +78,5 @@ type Backend interface {
 
 	GetRewardBase() common.Address
 
-	GetSubGroupSize() int
+	GetSubGroupSize() uint64
 }
